requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colorclass==2.2.2
docopt==0.6.2
Flask==3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
rfc3986==2.0.0
six==1.17.0
sniffio==1.3.1
spotipy==2.25.1
telebot==0.0.5
urllib3==2.4.0
Werkzeug==3.1.3
